Get rid of interactor
